refactor(web/chat): consolidate chat tool components into a single file
- Moved MainChatToolCheckIntent, MainChatToolDetectLanguage, MainChatToolGenerateFallback, MainChatToolHandleRequest, and MainChatToolWeather components into main-chat-render-tools.tsx for better organization.
- Removed individual tool files to streamline the codebase and improve maintainability.
- Updated import statements in the main chat component to reflect the new structure.
